Process arrow table for given query
moved array related functions to utils
