+++ src/cls/tabular/cls_tabular.h
@@ -14,11 +14,6 @@
 
 #include "include/types.h"
 
-#include <arrow/api.h>
-#include <arrow/io/memory.h>
-#include <arrow/ipc/writer.h>
-#include <arrow/ipc/reader.h>
-
 void cls_log_message(std::string msg, bool is_err, int log_level);
 
 #define STREAM_CAPACITY 1024
